requirements.txt
selenium~=4.1.3
numpy~=1.21.5
Pillow~=9.1.0